examples/period-app/python-backend/requirements.txt
# requirements.txt
# base requirements

-i https://pypi.org/simple
alembic==1.7.7
asgiref==3.4.1; python_version >= '3.6'
asyncpg==0.24.0
cffi==1.15.0
click==8.0.3; python_version >= '3.6'
# cryptography==36.0.2
databases[postgresql]==0.5.5
fastapi==0.75.0
h11==0.12.0; python_version >= '3.6'
httptools==0.4.0
markupsafe==2.0.1; python_version >= '3.6'
# psycopg2-binary==2.9.3
pydantic==1.8.2; python_full_version >= '3.6.1'
pyjwt==2.4.0
python-dotenv==0.20.0
python-editor==1.0.4
# pyyaml==6.0
sqlalchemy==1.4.32
starlette==0.17.1; python_version >= '3.6'
typing-extensions==4.1.1
uvicorn[standard]==0.17.6
uvloop==0.16.0
watchgod==0.8.1
websockets==10.2

anyio==3.3.4; python_full_version >= '3.6.2'
greenlet==1.1.2; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
idna==3.3; python_version >= '3.5'
# psycopg2-binary==2.9.3
pytest==7.1.1
# pyyaml==6.0
sniffio==1.2.0; python_version >= '3.5'
typing-extensions==4.1.1; python_version >= '3.6'
python-keycloak==0.27.0
opentdf==0.7.3
colorlog
aiohttp

boto3